+++ documentation/src/components/refine-week/refine-week-mobile.tsx
@@ -200,7 +200,7 @@ export const RefineWeekMobile = ({ variant, data, className }: Props) => {
                 <a
                   href={item.link}
                   target="_blank"
-                  rel="noreferrer"
+                  rel="noopener"
                   className={clsx(
                     "no-underline",
                     "flex items-center gap-2",