+++ documentation/src/refine-theme/common-github-star-button.tsx
@@ -15,7 +15,7 @@ export const CommonGithubStarButton = ({ className }: Props) => {
     <a
       href="https://github.com/refinedev/refine"
       target="_blank"
-      rel="noreferrer"
+      rel="noopener"
       className={clsx(
         "flex gap-2 items-center justify-center",
         "font-medium",